[DEVICE-LIST] - store and redux pattern for device list
