+++ src/store/device-types.ts
@@ -1,0 +1,3 @@
+export const GET_DEVICES = 'devices/GET_DEVICES';
+export const SET_DEVICES = 'devices/SET_DEVICES';
+export const FETCH_DEVICES = 'devices/FETCH_DEVICES';
