--- src/store/modules/devices/device-types.ts
@@ -1,1 +1,0 @@
-export const FETCH_DEVICES = 'devices/FETCH_DEVICES';
