+++ src/store/modules/devices/getters.ts
@@ -1,0 +1,7 @@
+import {GET_DEVICES} from "../../device-types";
+
+export const getters = {
+  [GET_DEVICES](state) {
+    return state.devices;
+  }
+};
