+++ src/store/modules/devices/index.ts
@@ -1,0 +1,14 @@
+import {mutations} from "../nav/mutations";
+import {actions} from "./actions";
+import {getters} from "./getters";
+
+const initialState = {
+  devices: [],
+};
+
+export default {
+  state: {...initialState},
+  mutations,
+  actions,
+  getters,
+}
