+++ src/store/modules/devices/mutations.ts
@@ -1,0 +1,9 @@
+import {SET_DEVICES} from "../../device-types";
+
+export const mutations = {
+  [SET_DEVICES](state, devices) {
+    state.devices = devices;
+  }
+};
+
+
