+++ src/store/modules/nav/actions.ts
@@ -1,0 +1,1 @@
+export const actions = {};
