+++ src/store/modules/nav/getters.ts
@@ -1,0 +1,2 @@
+export const getters = {
+};
